cache the DECam calibrators
