+++ tests/models/test_decam.py
@@ -304,7 +304,7 @@ def test_decam_download_and_commit_exposure( code_version, decam_raw_origin_expo
             #  downloads; this will always happen on github actions.
 
 
-@pytest.mark.skipif( os.getenv('RUN_SLOW_TESTS') is None, reason="Set RUN_SLOW_TESTS to run this test" )
+# @pytest.mark.skipif( os.getenv('RUN_SLOW_TESTS') is None, reason="Set RUN_SLOW_TESTS to run this test" )
 def test_get_default_calibrators( decam_default_calibrators ):
     sections, filters = decam_default_calibrators
     decam = get_instrument_instance( 'DECam' )
